# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful. OTP sent for verification.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid registration payload."
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "summary": "Register a new user account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP."
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "summary": "Verify account with OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/resend-otp": {
      "post": {
        "operationId": "AuthController_resendOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP resent successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Unable to resend OTP for the supplied email."
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "summary": "Resend account verification OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset OTP sent.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid email address."
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "summary": "Request password reset OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid reset payload or OTP."
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "summary": "Reset password using OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/privy": {
      "post": {
        "operationId": "AuthController_privySignIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivySignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session issued.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "summary": "Sign in with a Privy access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New session issued.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "summary": "Exchange a refresh token for a new session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Session revoked. No body."
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "summary": "Revoke a session",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ark Rides API",
    "description": "Ride-hailing backend for Ark Rides: rider and driver identity (Privy + email), ride lifecycle, fare ledger, driver wallet, emergency SOS and off-app booking channels.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "phone": {
            "type": "string",
            "example": "08012345678",
            "description": "Phone number with 10-15 digits"
          },
          "password": {
            "type": "string",
            "example": "strongpassword123"
          },
          "confirmPassword": {
            "type": "string",
            "example": "strongpassword123"
          },
          "acceptTerms": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "password",
          "confirmPassword",
          "acceptTerms"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ResendOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "strongpassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          },
          "newPassword": {
            "type": "string",
            "example": "newStrongPassword123"
          }
        },
        "required": [
          "email",
          "otp",
          "newPassword"
        ]
      },
      "PrivySignInDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Privy access token from the client SDK's getAccessToken().",
            "example": "eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "identityToken": {
            "type": "string",
            "description": "Privy identity token (the client's privy-id-token). Supplying it lets the server record the embedded wallet. Verified server-side, never trusted."
          },
          "audience": {
            "type": "string",
            "enum": [
              "rider",
              "driver"
            ],
            "example": "rider",
            "description": "Riders and drivers are separate accounts with separate id spaces, and one Privy DID may own one of each. The client states which it wants rather than the server guessing."
          },
          "name": {
            "type": "string",
            "example": "Amina Yusuf",
            "description": "Used only when provisioning a brand-new rider."
          }
        },
        "required": [
          "accessToken",
          "audience"
        ]
      },
      "RefreshSessionDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token returned by sign-in."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "ApiFieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "pickup.lat",
            "description": "Dotted path to the offending property."
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "lat must be between -90 and 90"
            ]
          }
        },
        "required": [
          "field",
          "messages"
        ]
      },
      "ApiFailureEnvelope": {
        "type": "object",
        "description": "Returned for every 4xx and 5xx. `message` is ALWAYS a single string; field-level detail goes in `errors`.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "integer",
            "example": 403
          },
          "message": {
            "type": "string",
            "example": "You can only view your own ride history."
          },
          "code": {
            "type": "string",
            "description": "Machine-readable discriminator; branch on this, not on the message.",
            "example": "FORBIDDEN",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "DUPLICATE_VALUE",
              "INVALID_REFERENCE",
              "MALFORMED_IDENTIFIER",
              "RATE_LIMITED",
              "DATABASE_ERROR",
              "INTERNAL_ERROR"
            ]
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiFieldError"
            },
            "description": "Present only for validation failures."
          },
          "path": {
            "type": "string",
            "example": "/api/v1/rides/abc"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "description": "Present only on 5xx. Quote it when reporting a fault."
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "code",
          "path",
          "timestamp"
        ]
      },
      "ApiSuccessEnvelope": {
        "type": "object",
        "description": "Returned for every 2xx that carries a body.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Request successful"
          },
          "data": {
            "description": "The payload. Its shape is per-endpoint.",
            "nullable": true
          },
          "meta": {
            "type": "object",
            "description": "Present on list endpoints.",
            "properties": {
              "page": {
                "type": "integer",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "example": 20
              },
              "total": {
                "type": "integer",
                "example": 41
              },
              "totalPages": {
                "type": "integer",
                "example": 3
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data",
          "timestamp"
        ]
      }
    }
  }
}